+++ 1037_div3.cpp
@@ -1,6 +1,6 @@
 
 // #include <bits/stdc++.h>
-// #define ll long long
+// #define ll ll
 // using namespace std;
 
 // int main() 
@@ -32,7 +32,7 @@
 
 
 // #include <bits/stdc++.h>
-// #define ll long long
+// #define ll ll
 // using namespace std;
 
 // int main() 
@@ -79,7 +79,7 @@
 
 // #include <bits/stdc++.h>
 // using namespace std;
-// using ll = long long;
+// using ll = ll;
 
 // int main()
 // {
@@ -124,7 +124,7 @@
 
 #include <bits/stdc++.h>
 using namespace std;
-using ll = long long;
+using ll = ll;
  
 int main() 
 {
@@ -177,7 +177,7 @@ int main()
 
 // #include <bits/stdc++.h>
 // using namespace std;
-// using ll = long long;
+// using ll = ll;
 
 // int main()
 // {
