+++ 873_div3A.cpp
@@ -6,7 +6,7 @@ int main()
     cin>>t;
     while(t--)
     {
-        long long int n;
+        ll int n;
         cin>>n;
         for(int i=1;i<=n;i++)
         {
