+++ 974_div3.cpp
@@ -1,5 +1,5 @@
 // #include <bits/stdc++.h>
-// #define ll long long int
+// #define ll ll int
 // using namespace std;
 
 // int main() 
@@ -40,7 +40,7 @@
 
 // #include <bits/stdc++.h>
 // using namespace std;
-// using ll = long long;
+// using ll = ll;
 
 // int main()
 // {
@@ -68,7 +68,7 @@
 
 #include <bits/stdc++.h>
 using namespace std;
-using ll = long long;
+using ll = ll;
 
 int main()
 {
