+++ 977_div2.cpp
@@ -1,6 +1,6 @@
 // #include <bits/stdc++.h>
 // using namespace std;
-// using ll = long long;
+// using ll = ll;
 
 // int main()
 // {
@@ -33,7 +33,7 @@
 
 #include <bits/stdc++.h>
 using namespace std;
-using ll = long long;
+using ll = ll;
 
 int main()
 {
