+++ 980_div2.cpp
@@ -1,5 +1,5 @@
 // #include <bits/stdc++.h>
-// using ll = long long;
+// using ll = ll;
 // using namespace std;
 
 // int main()
@@ -30,7 +30,7 @@
 // #include<bits/stdc++.h>
 // using namespace std;
 
-// #define ll long long
+// #define ll ll
 // #define pb push_back
 // #define all(x) x.begin(),x.end()
 // #define FastIO ios::sync_with_stdio(0);cin.tie(0); cout.tie(0);
@@ -76,7 +76,7 @@
 // }
 
 #include <bits/stdc++.h>
-using ll = long long;
+using ll = ll;
 using namespace std;
 int main()
 {
