+++ Equation_2000.cpp
@@ -1,5 +1,5 @@
 #include <bits/stdc++.h>
-using ll = long long;
+using ll = ll;
 using namespace std;
 
 int main()
